fix: mismatched types, missing error checks
arrayutils.[hc]:
binary_search now returns an ssize_t so that it can be negative.

input.[hc]:
get_int_input now returns a long, the same type that strtol returns.  It also properly checks whether strtol succeeded.  In the case that strtol failed, get_int_input now returns LONG_MIN, which the user is almost never going to use.

utils.h:
Since get_time_diff is only needed for Linux, time.h is now also only included on Linux.

wqunion.[hc]:
Instead of wastefully allocating a struct as well as two arrays for a total of three calls to malloc, init_wqunion_of_size now only allocates one array of WQUNION_NODEs, which pair an ID and size together.  It also properly checks whether malloc succeeded.

main.c:
Added more error checks as well as the ability to print a WQUNION.

+++ src/input.h
@@ -4,6 +4,6 @@
 // Keep the input retrieval process safe and robust
 // by using getnstr, which prevents the user from entering
 // more than the specified number of characters.
-int get_int_input(void);
+long get_int_input(void);
 
 #endif
